+++ judge-core/src/judge/mod.rs
@@ -1,8 +1,9 @@
-use crate::{compiler::Language, sandbox::ResourceLimitConfig};
+use crate::{compiler::Language, run::sandbox::RlimitConfigs};
 
 pub mod common;
 pub mod interact;
 
+#[derive(Debug, Clone)]
 pub struct JudgeConfig {
     pub language: Language,
     pub program_path: String,
@@ -11,5 +12,5 @@ pub struct JudgeConfig {
     pub output_file_path: String,
     pub answer_file_path: String,
     pub check_file_path: String,
-    pub rlimit_config: ResourceLimitConfig,
+    pub rlimit_configs: RlimitConfigs,
 }
