+++ judge-core/src/lib.rs
@@ -1,8 +1,7 @@
 pub mod builder;
 pub mod compiler;
 pub mod error;
-pub mod executor;
 pub mod judge;
 pub mod result;
-pub mod sandbox;
+pub mod run;
 pub mod utils;
